moved commands files
