+++ fast64_internal/oot/c_writer/oot_common_cmds.py
