+++ fast64_internal/oot/c_writer/oot_room_writer/oot_room_cmds.py
@@ -1,7 +1,7 @@
 from ...oot_level_classes import OOTRoom
 from ....utility import CData
 from ...oot_utility import indent
-from .oot_common_cmds import cmdAltHeaders, cmdEndMarker
+from ..oot_common_cmds import cmdAltHeaders, cmdEndMarker
 
 
 def cmdEchoSettings(room: OOTRoom):
